naksha-frontend/python/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
python-multipart==0.0.6 